tests/wpt/tests/css/css-shadow-parts/invalidation-complex-selector.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Invalidation Complex Selector</title>
    <meta href="mailto:fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="https://drafts.csswg.org/css-shadow-parts/" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <style>#elem #c-e::part(partp) { color: red; }</style>
    <script>installCustomElement("custom-element", "custom-element-template");</script>
    <template id="custom-element-template">
      <style>span { color: green; }</style>
      <span id="part" part="partp">This text</span>
    </template>
    The following text should be green:
    <div id="elem"><custom-element id="c-e"></custom-element></div>
    <script>
      "use strict";
      test(function() {
        const part = getElementByShadowIds(document, ["c-e", "part"]);
        const before = window.getComputedStyle(part).color;
        document.getElementById("elem").setAttribute("id", "new-elem");
        const after = window.getComputedStyle(part).color;
        assert_not_equals(before, after);
      }, "Part in selected host changed color");
    </script>
  </body>
</html>
